Move -sourcepath tests
Move -sourcepath tests so that the directory passed to
-sourcepath does not contain any unrelated files.
