+++ compiler/src/dotty/tools/dotc/core/SymbolLoaders.scala
@@ -131,7 +131,7 @@ object SymbolLoaders {
             ctx.warning(i"""$what ${tree.name} is in the wrong directory.
                            |It was declared to be in package ${path.reverse.mkString(".")}
                            |But it is found in directory     ${filePath.reverse.mkString(File.separator)}""",
-              tree.sourcePos)
+              tree.sourcePos.focus)
           ok
         }
