+++ tests/neg-custom-args/sourcepath/outer/nested/Test1.scala
@@ -1,4 +1,4 @@
-package completeFromSource.nested
+package outer.nested
 
 class Test4 {
