+++ tests/neg-custom-args/sourcepath2/hi/A.scala
@@ -1,0 +1,5 @@
+// Missing `package hi`
+
+class Hello { // error
+  val x: Int = 1
+}
