+++ tests/neg-custom-args/sourcepath2/hi/B.scala
@@ -1,0 +1,2 @@
+package hi
+class B
