+++ tests/pos-special/sourcepath/outer/Test2.scala
@@ -1,4 +1,4 @@
-package completeFromSource
+package outer
 import nested._
 
 class Test2 extends A(22) {
