+++ tests/pos-special/sourcepath/outer/nested/A.scala
@@ -1,4 +1,4 @@
-package completeFromSource.nested
+package outer.nested
 
 class A(y: Int) {
